+++ src/conn.ts
@@ -6,7 +6,7 @@ import { CODEFORCES } from "./parsers/codeforces";
  *
  * Util to create personal problems and debug this tool.
  */
-const PERSONAL = new SiteDescription(
+export const PERSONAL = new SiteDescription(
     "personal",
     "Not a site. Custom problems and contest.",
     "Contest name",
@@ -20,7 +20,7 @@ const PERSONAL = new SiteDescription(
             problems.push(new Problem(`P${i+1}`, `P${i+1}`, ["0\n", "2\n", "9\n"], ["2\n", "4\n", "11\n"]));
         }
 
-        return new Contest(problems);
+        return new Contest("personal", problems);
     },
     async problemId => {
         return new Problem(problemId, problemId, ["0\n", "2\n", "9\n"], ["2\n", "4\n", "11\n"]);
@@ -28,10 +28,44 @@ const PERSONAL = new SiteDescription(
 );
 
 /**
+ * Not a real site.
+ *
+ * Create an empty contest that will be filled by user manually.
+ */
+const EMPTY = new SiteDescription(
+    "empty",
+    "Not a site. Create empty problems",
+    "Contest name",
+    "Problem name",
+    async problemId => {
+        // Parse problemId. It is of the form problem-name-10
+        // Where `problem-name` is current name and `10` is number of problems
+        let args = problemId.split('-');
+
+        let numProblems =  args[args.length - 1];
+        let total = Number.parseInt(numProblems);
+
+        args.pop();
+        let name = args.join('-');
+
+        let problems = [];
+
+        for (let i = 0; i < total; i++) {
+            problems.push(new Problem(`P${i+1}`, `P${i+1}`, [], []));
+        }
+
+        return new Contest(name, problems);
+    },
+    async problemId => {
+        return new Problem(problemId, problemId, [], []);
+    }
+);
+
+/**
  * Register a new site creating an entry in this dictionary.
  */
 export const SITES: SiteDescription[] = [
-    PERSONAL,
+    EMPTY,
     CODEFORCES,
 ];